hook up hh_distc to hh_server
Summary:
This diff adds incrementallity support to hh_server after a hh_distc typecheck. This is done by overriding the base dep graph and instead use the one generated by hh_distc. This works since hh_distc is doing a full init in the background and captures all the deps necessary to construct a full dependency graph.

This diff noticeably doesn't clear the dep graph delta. This may result in slightly larger fanouts but not incorrect ones. That'll be cleared in the next few diffs.

Reviewed By: edwinsmith

Differential Revision: D44655968

fbshipit-source-id: a4e4c25e12d1e4a00496652314d58879ec137202