Add some tests
Summary:
Adds a basic test as well as examples of all the existing uses of hack_expr!, hack_stmt! and hack_stmts!.

Also modified the TokenStream test comparator to allow a slight difference with trailing commas (so that I can stop worrying about rustfmt changing the input in an incompatible way).

Reviewed By: edwinsmith

Differential Revision: D44505995

fbshipit-source-id: 0f87ab08b2c0438e31ba0e7153fe17b5ff8dd0c2