Do not add implicit supportdyn bounds for context constants
Summary: We add `as supportdyn<mixed>` to abstract type constants that lack bounds; this makes it so that `this::T` can be upcasted to dynamic by exposure to its supportdyn bound (whereas the trivial top type bound for abstract types cannot). However, this should not happen for context constants where we *do* expect to instantiate the constant with `[] === mixed`.

Reviewed By: periodic1236

Differential Revision: D44724859

fbshipit-source-id: 5a49129cfffeac0c7c3760deff78e368db434526