Fix handling of a backslash at the end of a line
Summary: One side effect of D44629355 is that a backslash at the end of a line in a multiline string is preserved while previously it was implementation-defined. This differs from the original intent (at least according to the old scarce documentation) of having the same semantics as in C++ where both the backslash and newline are omitted. Omitting is also consistent with Python and likely other languages so reintroduce it.

Reviewed By: prasad223

Differential Revision: D44729006

fbshipit-source-id: 8c181dcc5ff91ddbcec9f9bb8d7cc2367793b3cb