Make hack_expr! (and friends) hygienic.
Summary:
Added a trick to make the hack macros (somewhat) hygienic.

In a macro_rules macro you can use `$crate` to refer to the current crate but there's no way to do that in a proc_macro.  So switch to having the outer macro definition be a macro_rules macro that calls the proc_macro, passing `$crate`.

The bulk of this change is moving the content of the proc macro definitions to hack_macros_impl (with their own rust_library rule) and renaming them and then wrapping those in new macro_rules definitions.

Reviewed By: edwinsmith

Differential Revision: D44506000

fbshipit-source-id: 1c151341df78f716bb25ec7379485b5151e6b60f