Rewrite encoder
Summary:
The previous version of this code needed support from a `EmitTokens` trait which had to be implemented for each node in the AST telling it how to recursively turn that struct into a `TokenStream`.  That was expensive to maintain since it had to be manually updated.

I considered writing a derive macro to automatically build the `EmitTokens` trait - but then realized that it's ALMOST the same as `serde::Serialize` so I wrote an emitter that used `Serialize` (which the AST nodes already implement) to do the heavy lifting instead.

It's a little tricky because of two reasons:
(1) `Serialize` ignores transparent newtypes (like `Box`) so those have to be manually re-introduced.
(2) `Serialize` is EXTREMELY generic - so during serialization you can't look at the object being serialized in a specific way - you know you have a `T` but you can't ask if that `T` is an `Expr((), _, Expr_::String((_, &str)))` and handle that specially.  So instead you need to build that knowledge bottom-up (You see a `&str`, then later you realize that's part of a `(_, &str)`, then later you realize that's part fo a `Expr_::String((_, &str)))`, etc).

Most of the rewrite is isolated to the former `emit` module along with some fix-ups to deal with cleaner replacement description enums.  The macro parsing itself is untouched.

Reviewed By: shayne-fletcher

Differential Revision: D44505996

fbshipit-source-id: 33212cfcd6fa86e011c6293d238ef07e95cd70ba