Remove RepoOptions hash from cache breaker for facts
Summary: We've already removed the repo schema from the cache breaker for native facts, we want to do the same thing with the RepoOptions hash to avoid excessive cache invalidation between HHVM version bumps. We still need a mechanism to bump the cache key in the event of a breaking change to symbol parsing so I'm adding `Autoload.CacheBreaker` as a repo-option.

Reviewed By: jthemphill

Differential Revision: D44594560

fbshipit-source-id: b4785adea5851a40f1ddc32dc9eb4699febb6aa3