Fix Class Bounds for newtype super Constraints
Summary:
# Context

I noticed a bug with the experimental newtype_super_bounds feature. It would work if the super bound was a non-class type like `int` or `string`, but would not work if it was a class type. For example:

```
newtype T super C = mixed;

class C {}

// In another file

function f(C $c): T {
  return $c; // Errors
}
```

This diff intends to fix this.

# Mitigation
Examining `typing_subtype.ml` I determined the problem had to do with an earlier branch of a match. For a class type (`Typing_defs_core.Tclass`) we matched if it was an enum and then fell back to `default_subtype`. However this means we won't hit the code that handles super bound constraints in the default branch of the match.

The fix in this case was simple. I moved the check to be a guard and let the default case correctly handle the code.

Reviewed By: vassilmladenov

Differential Revision: D44765663

fbshipit-source-id: 2627c9d81fa65b296be4c56604146bfd8e3d41b4