Use an explicit try-catch rather than SCOPE_FAIL in a hot context
Summary: `SCOPE_FAIL` is fine for almost all use-cases, but in a hot context, the overhead of calling `std::uncaught_exceptions` to determine if an exception is currently being unwound can be excessive. Using an explicit try-catch-rethrow makes this slightly cheaper for the common case where no exception is being unwound.

Reviewed By: ricklavoie

Differential Revision: D44385988

fbshipit-source-id: 0a2ff8ccf8f4895fe9db455253c92ec481864e33