Use test runner temp path for embedded data instead of bin_root()
Summary:
bin_root() might be a buck-out path, which we really should
avoid writing into anymore. Also, if HHVM_BIN is set, it writes files
to where-ever it is, which is obnoxious. We already have a place to
store things like this (in the temp path the test runner creates), so
use that.

Reviewed By: mdko

Differential Revision: D44824257

fbshipit-source-id: 2a9235824b9b0f8e15bee8f8c8aa57f87b53251a