Simplify cached ekm impl
Summary:
Original version was not thread safe if called from worker threads. Instead of trying to grab locks, we can just pre-populate this like we do with the rest of transportInfo.
Simplify the cache to be read only and just be filled with the only known label values in our infra (this also makes sure the size of the cache will never grow unexpectedly

Reviewed By: knekritz

Differential Revision: D44590207

fbshipit-source-id: ded5974fb6b055cdb37123a8190cdaa5f0cdf3e8