log each shell-out-complete
Summary:
When the user invokes find-all-refs or hover or any other LSP message, then we log telemetry as CLIENT_LSP_METHOD.

I think it'd also be useful to have a separate CLIENT_LSP_SHELLOUT event for when we got an answer back to an "hh" shellout that we'd kicked off earlier. This event will arise a lot less frequently than CLIENT_LSP_METHOD, and can afford to store more information e.g. command-line.

That's what this diff provides.

Reviewed By: zhyty

Differential Revision: D44773503

fbshipit-source-id: 5f3ccdc7e56af02a3be2bd6570928c4a34795f0c