hook up hh_distc to hh_server
Summary:
This diff hooks up interrupts correctly so we can interrupt typechecks to hh_distc.

The correctness of this implementation relies on returning the whole files to check list back as "cancelled" if we experience any critical interrupts (eg. watchman telling us a file has changed on disk). This is in contrast to hh_server, which can incrementally update the dep graph and thus only return files that haven't been typechecked yet when interrupted. Because we can only update the dep graph all at once with hh_distc, we have to return the entire list of files to retypecheck if and when a hh_distc typecheck gets interrupted.

Additionally this diff refactors typing_check_service to take the root dir of www so we can have a root we can trust exists.

Reviewed By: ljw1004

Differential Revision: D44754667

fbshipit-source-id: 3df84b2ebf744bb5f75dd7c6025a71d45ac02588