Use iterator instead of reverse_iterator in iterator_class_is_container
Summary:
This was added back in D620607 in 2012. The code looked very different
then: it was checking if a class was a container by seeing if it could
be constructed with two parameters of a dummy class, and that diff
refined the check to the class being constructible from two iterator
parameters. reverse_iterator was the iterator chosen for that check.

This is causing problems now because libc++ enforces C++20's
requirement that a reverse_iterator can only be constructed from a
bidirectional iterator [1] via a static_assert. We have code which
attempts to instantiate iterator_class_is_container with an
unordered_map, and unordered_map::iterator is not bidirectional, so the
static_assert is fired and compilation fails.

Given the intent of the check, I believe that constructing a reverse
iterator is unnecessary, and just checking for constructibility with the
container's iterator will do, so this diff switches over to that.

[1] http://eel.is/c++draft/reverse.iterators#reverse.iter.requirements-1

Reviewed By: ispeters

Differential Revision: D44835336

fbshipit-source-id: a41cd6104407f99ec71dced4f2e36fe5f75879d9