Avoid reference to supportdynamic.hhi in errors
Summary: When instantiating the upper bound `as T` from `newtype supportdyn<+T> as T = ...`, the type-checker inserts reason information that manifests as a reference to `supportdynamic.hhi` in errors. This is especially odd when we elide `supportdyn`, and doesn't really add any useful information. So let's special-case instantiation of `supportdyn` to ensure that this doesn't happen.

Differential Revision: D44951888

fbshipit-source-id: 15f5b47ce0ce6c75292b9d65e3f42ad46c1cf996