Add current limit usage to CPU-CC EventHandler::onCycle callback
Summary:
Adds limitUsage (i.e. current active requests or QPS) to the CPUConcurrencyController::EventHandler::onCycle callback.

With this change, `getLimitUsage()` is always called once at the beginning of `cycleOnce()`. This ensures that the QPS tracking logic executed when running in TOKEN_BUCKET mode is refreshed every cycle (importantly, at the same cadence as the load is measured).

Reviewed By: joshkirstein

Differential Revision: D44652862

fbshipit-source-id: 77e6f1f3f3a5e07530d656c575b0fb0140dd05ad