Update Error Message for ExprTree Splicing
Summary:
Based on user feedback I am adjusting the error message for splicing that was added in D44517900. The feedback was to use less technical terms than splicing or lifting since first time users may not be aware what these terms meant.

No behavior change besides the text of the error message is changed in this diff.

Reviewed By: rjbailey

Differential Revision: D44971775

fbshipit-source-id: 8d745ac1b497656e4a89f123f90ad93e495e3f98