Fix a few bugs with default packages
Summary:
Code in the default package can be called by any other code, since it is included in deployments by default.

Code in a default package can still use if(package) syntax to call other code.

Also, change the error messages for default packages to be more clear; avoid using Pos.none and instead use the original module pos if the module does not belong to a package

Differential Revision: D44967770

fbshipit-source-id: 4395157f594723177ab102be12fcf11d94af8c73