Report static failure for SDT function calls not dynamic one
Summary: When we check a call to an SDT function, we check each argument in turn against the static parameter type and against a like-version of it. If both checks fail, we currently report the like-typed failure. This produces some poor error messages (e.g. involve `~_` where the `_` stands for a type variable). Better to report the static failure.

Differential Revision: D44964768

fbshipit-source-id: 5821a254069432551aaba89d5543e2afdd5562fc