6/7 DocumentSymbol doesn't need location
Summary:
I wish to make some systematic changes to clientIdeMessage.ml in future, and in preparation I'm spending some time streamlining it.

`textDocument/documentSymbol` is a request to fetch all symbol definitions in a file, so they can populate the Cmd+Shift+O "Go To Symbol In Editor".

```
  | Document_symbol : document * location -> FileOutline.outline t
```

It doesn't make sense that this request should take a `location` (i.e. line+column). This diff therefore removes it:
```
  | Document_symbol : document -> FileOutline.outline t
```

Reviewed By: nt591

Differential Revision: D44991597

fbshipit-source-id: 3914ff902c4e40e9ddafff8e8f3ae1dfbdcbf1ca