powered_by:Serverless_ide
Summary:
This is minor cleanup. The function `publish_errors_if_standalone` is used to publish diagnostics that we got back from clientIdeDaemon. Therefore it will always be `~powered_by:Serverless_ide`. Just makes the code a touch neater.

(in preparation for a future diff when 15 other methods will also need to publish clientIdeDaemon errors as well)

Reviewed By: nt591

Differential Revision: D44985945

fbshipit-source-id: 73e716064ad9b555d52c5d13f74f1be007162fc7