Add cross package to typing defs::fun_type
Summary: Adds cross package attribute to the function type. This will allow us to then implement subtyping for the attribute later. This diff does not add any population, and the field will be None at all times until a diff above the stack populates the fields in declaration and elsewhere.

Differential Revision: D44879312

fbshipit-source-id: 69fd53ea042df036521856697e52335245459437