Case Type Subtyping
Summary:
# Context
This diff contains the changes to subtyping required to support case types. In general there are no changes to be made when case types appear in a subtype position. They are encoded as newtypes with upperbounds, so the existing subtyping logic applies.

There are some changes needed for when a case type appears in a supertype position. The typing rule is:

```
(case type CT = V) ∈ Γ
Γ ⊢ T <: V
--------------------
Γ ⊢ T <: CT
```

Stated in english, a type `T` is a subtype of `CT` if `CT` is a name of a case type with variant type `V` and `T` is a subtype of `V`.

This is the same logic used for the experimental newtype_super_bounds feature.

# Implementation
When implementing this rule I had to fix it so that we properly substituted type params when localizing the lower bound type. Without this we would fail to type check case types that utilize generics, i.e.

```
case type CT<T> = vec<T>;

vec<int> <: CT<int>
```
Without this change the above subtyping judgement would not hold.

The other thing I needed to do was decompose some of our built-in unions to their union parts. This is done already for `num` (line 1694 in Typing_subtype) earlier in simplify subtype. This however isn't done for `arraykey`. This means that if we have:

```
case type CT = int | string;

arraykey <: CT
```

would not hold. I elected to do the decomposition only when dealing with newtypes, but if this should be applied earlier I can move it earlier in simplify_subtype next to where we do the decomposition for `num`. I also found the need to do the same for nullable types, in particular `?arraykey`.

# Design Note
We can potentially generalize this kind of decomposition when deciding subtype relations between two case types. I.e allowing:

```
case type SubCT = T1 | T2;

case type SuperCT = T1 | T2 | T3;

SubCT <: SuperCt
```
It isn't immediately clear to me if this is desirable since it breaks the opaqueness of `SubCT`. If this kind of relation is desired, it can be achieved by adding a bound to `SubCT`, i.e. `case type SubCT: SuperCT`. If we get feedback from users that this kind of decomposition is useful in general we can revisit.

This option is not available for our builtin union types, thus the decomposition is necessary.

Reviewed By: mpu

Differential Revision: D44811624

fbshipit-source-id: 3210864408c204f98749c4897cc2ac0cf4b29b53