Case Type Direct Decl Parser
Summary:
This implements the direct decl parser from FFP to AAST for the case type feature.

This should produce the same output detailed in D44811609

The implementation is based `make_alias_declaration`. The other thing I needed to change was `make_case_type_variant` to return the type, since we do not encode the list of variants using `ListItem`.

Reviewed By: rjbailey

Differential Revision: D44811612

fbshipit-source-id: 93fa7b3b8bf120a098172415e45952cb917656e6