Case Type Parser Errors
Summary:
This implements the parser errors for the case type feature. See D44810190 for description of desired parsing.

We should report a subset of the errors reported by `AliasDeclaration`. Specifically:
* check_attr_enabled()
* invalid_modifiers_errors()
* check_type_name()
* errors::type_alias_to_type_constant

There is a decent amount of duplication here. I believe the best way to eliminate it would be to move the majority of the checks from being done on the ffp to be done on the lowered AST. When rust_parser_errors was introduced we didn't have any checks done on the lowered ast that ran for both HackC and hh, but now that we are adding elaboration passes on the lowered ast we could move those checks to be done during that pass.

That is a larger change than I want to commit to right now, so I rather leave as is for now and do the clean up at another point.

This diff also adds unstable feature gating and thus is the first part of the stack I feel comfortable landing at this point. This with the previous diffs in the stack I plan to land together.

Reviewed By: rjbailey

Differential Revision: D44811601

fbshipit-source-id: 23c6d7f759b0fa1e71f75531464bd2f3e201a61a