Case Type Lowerer
Summary:
This implements the lowering from FFP to AAST for the case type feature.

We will encode case types as type aliases. The lowering will work mainly as follows. For a case type:
```
case type CT as U1, U2, Un = T1 | T2 | Tn;
```

It will be lowered into the following form:
```
newtype CT as (U1 & U2 & Un) = (T1 | T2 | Tn);
```

Only other special cases are the following:
```
case type CT as U = T;
```
is lowered to

```
newtype CT as U = T;
```

Specifically if there is only a single upper bound or a single variant then we don't bother to wrap in an intersection/union type.

Finally if we somehow get a FFP that doesn't have any variants we will report a `missing_syntax(...)` error since should be reported as an error by the parser.

Reviewed By: rjbailey

Differential Revision: D44811609

fbshipit-source-id: 4cdb7f1c20bd8ec19464f8cb48f5791a67d79cbb