Case Type FFP Schema
Summary:
Case types will be a way of denoting a subset of union types with particular restrictions that allow them to be treated as ADTs based on the tags that HHVM already carries for values. This diff is the first in a stack that introduces parser support for the feature, containing the changes to the full fidelity parser schema.

It is very similar to the FFP schema for `AliasDeclaration` with a `case` modifier appearing before `type` (i.e. `case type Id`). We also do not support he `module` modifier. Other differences in the grammar is supporting declaring multiple upper bounds using `as` and a comma separated list of types. Finally to the right of the `=` it will contain a list of types separated by `|`. A typical case type will then be written as:

```
case type Ident as U1, U2 = T1 | T2 | Tn;
```
Alternatives considered for the grammar was using something resembling enums i.e.
```
enum type Ident: U1, U2 {
  case T1;
  case T2;
  case Tn;
}
```

This felt more verbose and the representation we will use in the runtime will be closer to an alias than an enum.

One choice in the schema that is relevant is my decision to include a new node `CaseTypeVariant` instead of using `Specifier`. In the future I imagine we will want to support more features for the variants of a case type i.e. allowing attributes per variant or having them define class/data types inline which will need to support a larger grammar than `Specifier`.

Note: No effort was made to introduce decent formatting for hack_format. This will be done before the feature enters preview.

Reviewed By: rjbailey

Differential Revision: D44810190

fbshipit-source-id: a09164208c2feeeebe1b7b5f7cabb6f30d646325