Introduce TypedefVisibility::CaseType
Summary:
# Context

Up to this point, case types are treated as newtypes. However there are some instances when we want to treat case types differently from a newtype in the type checker. In particular for a case type they never act transparently. This is different from newtype who act transparently within a file. In particular the following is valid for a newtype:
```
newtype NT = int;

function f(NT $x): int { return $x; }
```

While this would be rejected for a case type:
```
newtype CT = int;

function f(CT $x): int {
  return $x; // Error
}
```

# Implementation

To be able to distinguish a newtype from a case type, I added a new kind of `TypedefVisibility`. This allows us to change the behavior of case types for `Typing_env.is_typedef_visible`. For case types this will always return false, resulting in creating a `Tnewtype` whenever a case type is localized.

One thing I took care of was auditing all the current places where we force expand aliases. There were two locations that were relevant:
* Typing_memoize - this performs a check to see if the parameter is a valid parameter type for a memoized function
* Typing_equality_check - this is a linter that checks if `===` between types would always produce false

While ideally we would never expand a case type, I elected to allow them to be expanded since these are only for checks and the expanded forum will not leak into the TAST or impact inference.

Reviewed By: shayne-fletcher

Differential Revision: D44811617

fbshipit-source-id: 9f76090289149f4c1dba0ee36186b52ed222ffe7