streamline HackEventLogger.invariant_violation_bug
Summary:
I think it's good practice to use [HackEventLogger.invariant_violation_bug]. This diff makes the function a touch easier to use.
* path, pos, telemetry are now optional
* the key [~desc] is no longer a named parameter
* I added two optional parameters `[data]` and `[data_int]` to make it easier to put extra strings.

BEFORE:
```
    HackEventLogger.invariant_violation_bug
      ~desc
      ~path:Relative_path.default
      ~pos:""
      (Telemetry.create ());
```

AFTER:
```
HackEventLogger.invariant_violation_bug desc ~telemetry;
```

BEFORE: This example wasn't good practice because it had such a long string in the key, hence harder to aggregate in scuba. I hope that by having a convenient [~data] field, then people will be less inclined to put stuff in the key.
```
          HackEventLogger.invariant_violation_bug
            ~desc:
              ("nastInitCheck can't find expected class prop. class_name = "
              ^ class_name
              ^ "; member_origin = "
              ^ member_origin
              ^ "; prop_name = "
              ^ prop_name)
            ~path:Relative_path.default
            ~pos:""
            (Telemetry.create ());
```

BEFORE: This example wasn't good practice because the msg (1) affected the key, (2) was duplicated in telemetry. Again I'm hoping that the [~data] field will make it easier to avoid this.
```
    HackEventLogger.invariant_violation_bug
      ~path:Relative_path.default
      ~pos:""
      ~desc:("Unexpected hh_distc error: " ^ msg)
      (Telemetry.create () |> Telemetry.string_ ~key:"error" ~value:msg);
```

It would have been nice to make the key be a named parameter `~key:"unexpected distc error"` but that's not possible because all the other parameters are optional.

Reviewed By: dlreeves

Differential Revision: D45008162

fbshipit-source-id: e03a381a8d52b44aa28788f5aaf92a3730a83eb4