Standardize exit codes
Summary:
We use a variety of exit codes in an inconsistent manner. Standardize
on generally using HPHP_EXIT_FAILURE for error, and 0 for success. Use
1 for a few situations, like for a script failing with a fatal.

This cleanup helps me improve the error recognition in hphp/test/run
as I can more reliably interpret the exit codes (which is not really
possible right now).

Reviewed By: oulgen

Differential Revision: D45029897

fbshipit-source-id: f0b96f161073e8f493b77bb0d1ad33e2bc827e9d