Add Tag Closing for XHP Autocompletion
Summary:
D44310250 added the autocompletion for required attributes which developers need to always include. We also need to include tag closing depending on whether tag needs to contain children or not. Previously, this couldn't be done easily because there was no children information available through the data structure. Thank to D44691847, we can now learn if the children of XHP class is empty or not.

This diff, per title, adds closing tag for XHP based on children empty marked.

Reviewed By: antonk52

Differential Revision: D45057543

fbshipit-source-id: 86a4be324f8e79fad57daa0922cca05fc9b7467e