hhvm enforcement using hhbbc-time original module name resolution
Summary:
With "module level traits" methods of public traits are considered part of the module they are defined in, instead of the module they are copied into.  This allows them to call internal top-level functions of the module they were defined in, allowing developers to export internal methods in a controlled way.

This diff proposes a prototype implementation of the HHVM revised enforcement.  Contrarily to D43186249 where the module name of a method is resolved at run-time, this diff computes the original module name when HHBBC builds the repo, and stores it explicitly in the Func representation.  This is not done for all methods (overhead is considered too high), but only for methods defined in a trait the is defined in a module but used by a class that belongs to a different module.  In practice:

- HHBBC representation of methods is extended with the name of the module where the method was originally defined (irrespective of trait inlining), and a flag `requiresFromOriginalModule` that is set to true when HHBBC detects that a trait defined in module A is being inlined in class defined in module B.   This is performed when methods are cloned, reusing the existing distributed machinery (since Units are not available to the distributed machinery, the HHBBC representation of a Class is extended with the module where it was defined).

- if HHVM loads a method with the `requiresFromOriginalModule` field set,  then it allocates `extendedSharedData` for the method and stores the original module name in the _new_ field  `originalModuleName` of the `extendedSharedData`;
- HHVM `Func::moduleName` returns the original module name if the field is defined, otherwise it returns the standard `unit()->moduleName()` value.

If code is executed in sandbox mode, then the module name obtained via `unit()->module()` is already accurate for methods defined in trait T in module A and used in class C in module B wrt the module level trait semantics.

As a result, memory overhead to store the original module name only occurs in repo mode for methods defined in a trait used by a class in a different module.

Traits must explicitly opt-in in the "module level traits" semantics, using the trait attribute `<<__ModuleLevelTrait>>`.

Reviewed By: ricklavoie

Differential Revision: D44010949

fbshipit-source-id: 53e1b592e497a2ee64bb332c477eee54694a9220