telemetry to remove "default_path"
Summary:
When we first wrote clientLsp, there were some irritating cases where hh_server returned an empty path for various RPC calls over the persistent connection (hover, error-squiggles, I can't remember what). This was a pervasive problem so I had addressed it with a fallback in clientLsp: "if we get an empty path, then use this `default_path` instead."

I regret that now! Also, I don't know if it's even still needed. And if it is, I'd rather address it at the source in hh_server.

This diff adds telemetry if the fallback is ever used.

Reviewed By: Wilfred

Differential Revision: D45121703

fbshipit-source-id: 11bdf97813c5a6321f7ef13346125035a7567e59