cache all_errors rather than recomputing each time
Summary:
Before this diff, whenever we fetched errors for a Provider_context.entry, we would dynamically merge the ast_errors with the naming_and_typing_errors. This diff changes it to merge them just once.

Why am I doing this? Because in future diffs I want to be able to detect when the errors-list for an entry has been recomputed, so I know whether to publish updated diagnostics for the file. I figured a good way to do that is for clientIdeDaemon to maintain a list of all the latest errors it has published, and then it can compare using physical equality to know whether an entry's error-list is newer.

Reviewed By: Wilfred

Differential Revision: D45100837

fbshipit-source-id: 7695810ab7240b22449ae91de0730d2023721dae