more explicit state monad
Summary:
This is some mindless refactoring. Previously, each handler in clientIdeDaemon would match on receiving a message while in the `Initialized` state (meaning, it had a saved-state loaded and was working correctly), but it would hide that information and not know what state the handler left it in.

This diff adjusts the code to make explicit that it remains in the Initialized state after handling a message.

Why am I doing this? I don't actually need to! But I encountered the need during a (since abandoned) version of diffs on top of this stack, and it seemed like a clean improvement, so I went with it.

Reviewed By: Wilfred

Differential Revision: D45111176

fbshipit-source-id: fd0b78467b8395ba86f5c88d362c21c105aed77e