Include type of unknown fields in shape types
Summary:
This is a change to the internal representation of `shape` types. Instead of simply recording whether a shape is *closed* or *open*, we record the *type* of the unknown fields in an open shape. This lets us represent open shapes whose unknown fields support dynamic type (i.e. are of type `supportdyn<mixed>`) whilst allowing known fields to be of non-supportdyn type. An ordinary open shape type has type `mixed` for the unknown fields.

Open shape types that have non-mixed unknown fields are not denotable in Hack syntax, but are pretty-printed as e.g. `shape('a' => int, _ => nonnull)`. Here the type of unknown fields is `nonnull`. So `shape('a' => int, ...)` is equivalent to `shape('a' => int, _ => mixed)`.

Future simplification: use `nothing` for the type of unknown fields to represent closed types.

Reviewed By: vassilmladenov

Differential Revision: D45001642

fbshipit-source-id: 973af2f97df44dfc4132f6811eece3c95f1a0343