housekeeping: remove --split-hphpc
Summary:
No automation uses this parameter, and the only way it would work is if you set
HHVM_BIN and manually place a binary there called `hphpc`. AFAIK this
would be an undocumented workflow.

There are no buck targets that create a `hphpc` binary, and even if HHVM_BIN
points to a buck-out path for `hhvm`, that will be the wrong path under buck2
and it's a code smell to assume anything about buck-out directory layout.

Reviewed By: ricklavoie

Differential Revision: D45211689

fbshipit-source-id: bb41ffc246bf88be22ee0d274b42a9097bb387d3