Remove nix from github actions (#9359)
Summary:
This has been failing since February; we have passing Ubuntu builds on GitHub actions now.

One difference is that the nix build runs the quick hhvm tests, but Ubuntu doesn't; while we'd like to get the tests passing on Ubuntu, I'm not seeing it as a practical issue, given the nix failure is a build failure - it doesn't get to the tests.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9359

Reviewed By: Wilfred

Differential Revision: D45236769

fbshipit-source-id: 05727a100a448cdb036270749f51677f1480f1a9