Handle SetS to pull in constants
Summary:
HHBC has no notion of a direct lookup version of GetClass<string> or SetS<string::string> - it always has to go through an indirect constant.

When lowering to textual translate a static function's SetS<constant::constant> to use a built-in.

We'll probably want to eventually lower this to a SIL builtin like
```
SetStaticProp(<CLASS>, "$prop", value)
```
but for now this will do.

Reviewed By: davidpichardie

Differential Revision: D45073780

fbshipit-source-id: 3278024c0e7a3fe44eec88ebc6673c071ae1bd14