Thread-cached RequestContext data lookup
Summary:
`RequestData` retrieval involves a concurrent hashtable lookups, which can be expensive if it happens on a hot path. The request context is a natural choice to deliver per-request data down deep stacks, so it is desirable to make access fast.

This diff adds a thread-cached version of `getContextData()`. The result is cached in a thread-local variable, and implicitly invalidated every time the current request context changes, or it is modified.

Reviewed By: blezsan

Differential Revision: D45151135

fbshipit-source-id: 29a6abe8c7c26aa13a5a589c65f1b6fd6dce3396