Improve Formatting for Case Types
Summary:
The initial formatting support for case types was minimal, but not ideal. This diff implements enough formatting which I believe is good enough for an initial launch.

The behavior of the formatting changes depending on if the variant has a leading bar (`|`) token. If it does then the variants will always be split on new lines, even if they could fit on a single line. Example is:

```
case type CT = | T1 | T2;
```
Would turn into
```
case type CT =
  | T1
  | T2;
```

If there is no leading bar, then we will attempt to fit it on a single line, i.e.
```
case type CT = T1 | T2;
```
The only thing that is less than ideal is if there is no leading bar, but it cannot fit on a single line. In that case the formatting will be
```
case type CT = T1
  | T2
  | ...
  | Tn;
```
Ideally this would be:
```
case type CT =
  | T1
  | T2
  | ...
  | Tn;
```

But inserting a new token is not supported in general for the formatter.

Reviewed By: rjbailey

Differential Revision: D45280758

fbshipit-source-id: 348ae71bbefeba8269a53e1091c47291486f919d