use Rust batch indexer to process file changes after saved state load
Summary:
The daemon will process all file changes between the naming table saved state load and the current state of the repo at once before responding to further requests. For performance reasons, we process these changes in Rust in multiple threads.

We're choosing to batch process all of these file changes immediately because we're allowing the daemon to generate errors on its own, and we'd like to avoid showing stale errors out of the gate.

This feature is gated by `ide_process_changes_sync`.

Note: eventually, we want to use the multithreaded batch indexer to process all file changes.

Reviewed By: ljw1004

Differential Revision: D44595899

fbshipit-source-id: 9e0172ee2fa04ecbb125cc386120e119aab8c66b