refactor out OSS parts of hh24_test
Summary:
We introduce a dependency to `hackrs_provider_backend` in a later diff. This causes problems with the OSS build because `hackrs_provider_backend` has a test dependency on `hh24_test`, which has its own dependencies that interfere with our OSS build. Unfortunately, it seems that we can't make the distinction between test-dependency and build-dependency in OSS yet (?).

This diff we make `hh24_test` mostly OSS -- some of its dependencies don't necessarily need to be internal, but will take some more work to move out.

Reviewed By: CatherineGasnier

Differential Revision: D45248985

fbshipit-source-id: 89273bba7bad9dfc645038044b73b4560310ffbb