Fix Rust service name in the context of interactions
Summary:
In Rust and C++, Thrift interactions are implemented as sub-services under the
parent service, where the parent service is the one defined in the spec with
`service <service name>`. This is why `service:name` refers to the interaction
name when generating code within the context of an interaction.

Currently, the templates for Rust code use `service:name` in a few locations.
Many of those are expected because the interaction is reusing the template for
"standard" services. However, to match generated C++ code, there are a few
locations that should use the parent service name. This diff modifies the
service name used when creating a context stack to be the parent service name
to match C++ when in the context of an interaction.

Reviewed By: jsgf

Differential Revision: D45252447

fbshipit-source-id: 63219a720aa0648f4af6cd2608839dc12b40f25b