Introduce a bias when checking `T < ~U` towards the inner error
Summary:
For `T <: (U | ... | V)` we have a special case that says if you can strip a dynamic out of the union and get a singleton type `W`, we do `T <: W ||| T <: dynamic`, but our typing error represented by `~fail` is still pointing to the outer types. This introduces a bias towards the result of `T <: W` if we find that `T <: dynamic` has failed.

This is morally similar to {D44964768}

Reviewed By: andrewjkennedy

Differential Revision: D45306025

fbshipit-source-id: c6d760823437e503abff2702d0727427f8d621c1